Start with rewrite in Haskell.
Removes all of the Python stuff and starts with just a small DMX test
wire (we're using netwire now for DMX output). Also, we're already
including SDL 2 as a dependency, because we're going to use it for the
interface (no curses based display anymore, because it's too much
overhead).

The main reason I'm doing this is because the original project in Python
was too troublesome to maintain. And as the machine where it was running
broke quite badly (CPU overheat, memory errors), it's really time to
throw away the old implementation. Especially because with all that
statefulness combined with threading we were "able" to constantly keep a
Core 2 Duo at 100% CPU load on both cores.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>