Implement blitting of CP437 chars and strings.
We're now also tying in a preliminary SDL main loop, just to see "Hello
World!" shining in bright white :-)

Of course, all operations here are done using lists, and we might want
to do better if we'd use maps. But we need to get things working first
and we can still optimize later.

And as you probably see quite instantly in the diff: Yes, that's a part
of CP437 which I have "apic-quasiquoted" into the source :-D

Signed-off-by: aszlig <aszlig@redmoonstudios.org>