UI.Sprite: Prepare for nested sprites.
The reason I'm doing this at the most nested part is that we can
recursively emit subsprite data and afterwards we can flatten the whole
data in one go. One advantage of this approach is that it is easy to
add constructors for functions, so that we can render sprites
dynamically based on the function application.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>