Add a few sprites for experimenting with nesting.
In the end we're going to support revos and LED sticks, so this are a
few sprites that we're going to use. Eventually we're going to add wires
for nested sprites, so we might end up creating separate textures in GPU
memory.

But so far and as I mentioned in the previous commit, it's for
experimentation purposes.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>