Move over to new Rockfabrik stage setup.
Yep, that's a commit you wouldn't expect from me at all, but I'm a big
douchebag and accidentally rm -rf`ed ~/granpa, which is my local working
directory with all the changes I've made within the last few hours.

While I haven't lost the source code (phew), I lost all the commits and
this is a summary of the changes this commit squashes together:

 * Change DMX addresses of the stage setup.
 * Change visual appeareance of the viewport.
 * Don't center the clock among the Y axis.
 * Add four additional flash keys.
 * Add two keys to select groups (of rigs).

That's what I remember so far and I unfortunately don't have the time to
git add -p all of it, trying to reproduce the commits I had before.

Sorry for that.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>