inputdev/mouse: Actually use the timeout value.
Having select without a timeout means that the mouse handler doesn't
proceed or even stop whenever there is no input.

This led to a bug that whenever we're going to quit the application,
it simply wasn't quitting until you actually moved the mouse.

Now it should quit (or .join()) after one second.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>